requirements.txt
Flask==3.0.0
biopython==1.81
lxml==5.1.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
email-validator==2.1.0.post1
pytz